FIX:[DEV-10561] update bar labels with CI (#1973)
* [DEV-10561] update bar labels with CI

* [DEV-10561] Add Story for CI bars with labels
